polish utility component

+++ tests/utility/summation.cpp
@@ -14,18 +14,14 @@
 
 
 
-// Function to compute factorial
-inline static long double factorial(int n) {
+inline static long double factorial( int n ) 
+{
     if (n <= 1) return 1;
     return n * factorial(n - 1);
 }
 
-
-
-
-
 // Function to compute the determinant of the Hilbert matrix
-inline static double hilbert_determinant(int n) {
+inline static long double hilbert_determinant(int n) {
     
     long double detinv = 1.;
 
@@ -39,7 +35,7 @@ inline static double hilbert_determinant(int n) {
     
     std::cout << "H1: " << 1./(double)detinv << '\n';
     // The determinant is c_n / denominator
-    return 1./static_cast<double>(detinv);
+    return 1./detinv;
 }
